Generate a separate file for module_graph
