+++ core/bessctl.cc
@@ -50,6 +50,7 @@
 #include "message.h"
 #include "metadata.h"
 #include "module.h"
+#include "module_graph.h"
 #include "opts.h"
 #include "packet.h"
 #include "port.h"
@@ -723,7 +724,7 @@ class BESSControlImpl final : public BESSControl::Service {
     }
 
     // Check local constraints
-    for (const auto& pair : ModuleGraph::all_modules()) {
+    for (const auto& pair : ModuleGraph::GetAllModules()) {
       const Module* m = pair.second;
       auto ret = m->CheckModuleConstraints();
       if (ret != CHECK_OK) {
@@ -1099,7 +1100,7 @@ class BESSControlImpl final : public BESSControl::Service {
 
   Status ListModules(ServerContext*, const EmptyRequest*,
                      ListModulesResponse* response) override {
-    for (const auto& pair : ModuleGraph::all_modules()) {
+    for (const auto& pair : ModuleGraph::GetAllModules()) {
       const Module* m = pair.second;
       ListModulesResponse_Module* module = response->add_modules();
 
@@ -1129,8 +1130,8 @@ class BESSControlImpl final : public BESSControl::Service {
 
     std::string mod_name;
     if (request->name().length()) {
-      const auto& it2 = ModuleGraph::all_modules().find(request->name());
-      if (it2 != ModuleGraph::all_modules().end()) {
+      const auto& it2 = ModuleGraph::GetAllModules().find(request->name());
+      if (it2 != ModuleGraph::GetAllModules().end()) {
         return return_with_error(response, EEXIST, "Module %s exists",
                                  request->name().c_str());
       }
@@ -1161,8 +1162,8 @@ class BESSControlImpl final : public BESSControl::Service {
                                "Argument must be a name in str");
     m_name = request->name().c_str();
 
-    const auto& it = ModuleGraph::all_modules().find(request->name());
-    if (it == ModuleGraph::all_modules().end()) {
+    const auto& it = ModuleGraph::GetAllModules().find(request->name());
+    if (it == ModuleGraph::GetAllModules().end()) {
       return return_with_error(response, ENOENT, "No module '%s' found",
                                m_name);
     }
@@ -1183,8 +1184,8 @@ class BESSControlImpl final : public BESSControl::Service {
                                "Argument must be a name in str");
     m_name = request->name().c_str();
 
-    const auto& it = ModuleGraph::all_modules().find(request->name());
-    if (it == ModuleGraph::all_modules().end()) {
+    const auto& it = ModuleGraph::GetAllModules().find(request->name());
+    if (it == ModuleGraph::GetAllModules().end()) {
       return return_with_error(response, ENOENT, "No module '%s' found",
                                m_name);
     }
@@ -1224,16 +1225,16 @@ class BESSControlImpl final : public BESSControl::Service {
     if (!m1_name || !m2_name)
       return return_with_error(response, EINVAL, "Missing 'm1' or 'm2' field");
 
-    const auto& it1 = ModuleGraph::all_modules().find(request->m1());
-    if (it1 == ModuleGraph::all_modules().end()) {
+    const auto& it1 = ModuleGraph::GetAllModules().find(request->m1());
+    if (it1 == ModuleGraph::GetAllModules().end()) {
       return return_with_error(response, ENOENT, "No module '%s' found",
                                m1_name);
     }
 
     m1 = it1->second;
 
-    const auto& it2 = ModuleGraph::all_modules().find(request->m2());
-    if (it2 == ModuleGraph::all_modules().end()) {
+    const auto& it2 = ModuleGraph::GetAllModules().find(request->m2());
+    if (it2 == ModuleGraph::GetAllModules().end()) {
       return return_with_error(response, ENOENT, "No module '%s' found",
                                m2_name);
     }
@@ -1271,8 +1272,8 @@ class BESSControlImpl final : public BESSControl::Service {
     if (!request->name().length())
       return return_with_error(response, EINVAL, "Missing 'name' field");
 
-    const auto& it = ModuleGraph::all_modules().find(request->name());
-    if (it == ModuleGraph::all_modules().end()) {
+    const auto& it = ModuleGraph::GetAllModules().find(request->name());
+    if (it == ModuleGraph::GetAllModules().end()) {
       return return_with_error(response, ENOENT, "No module '%s' found",
                                m_name);
     }
@@ -1343,7 +1344,7 @@ class BESSControlImpl final : public BESSControl::Service {
 
     if (request->module_name().length() == 0) {
       // Install this hook on all modules
-      for (const auto& it : ModuleGraph::all_modules()) {
+      for (const auto& it : ModuleGraph::GetAllModules()) {
         if (request->enable()) {
           *response =
               enable_hook_for_module(it.second, gate_idx, is_igate, use_gate,
@@ -1360,8 +1361,8 @@ class BESSControlImpl final : public BESSControl::Service {
     }
 
     // Install this hook on the specified module
-    const auto& it = ModuleGraph::all_modules().find(request->module_name());
-    if (it == ModuleGraph::all_modules().end()) {
+    const auto& it = ModuleGraph::GetAllModules().find(request->module_name());
+    if (it == ModuleGraph::GetAllModules().end()) {
       return return_with_error(response, ENOENT, "No module '%s' found",
                                request->module_name().c_str());
     }
@@ -1467,8 +1468,8 @@ class BESSControlImpl final : public BESSControl::Service {
       return return_with_error(response, EINVAL,
                                "Missing module name field 'name'");
     }
-    const auto& it = ModuleGraph::all_modules().find(request->name());
-    if (it == ModuleGraph::all_modules().end()) {
+    const auto& it = ModuleGraph::GetAllModules().find(request->name());
+    if (it == ModuleGraph::GetAllModules().end()) {
       return return_with_error(response, ENOENT, "No module '%s' found",
                                request->name().c_str());
     }
@@ -1581,8 +1582,8 @@ class BESSControlImpl final : public BESSControl::Service {
       c = it->second;
     } else if (class_.leaf_module_name().length() != 0) {
       const std::string& module_name = class_.leaf_module_name();
-      const auto& it = ModuleGraph::all_modules().find(module_name);
-      if (it == ModuleGraph::all_modules().end()) {
+      const auto& it = ModuleGraph::GetAllModules().find(module_name);
+      if (it == ModuleGraph::GetAllModules().end()) {
         return_with_error(response, ENOENT, "No module '%s' found",
                           module_name.c_str());
         return nullptr;
